requirements.txt
selenium-requests==1.3
selenium==3.13.0
requests==2.19.1
mistune==0.8.3
matrix-bot-api==0.1
matrix-client==0.3.2
lxml==4.2.3
Cython==0.28.4
beautifulsoup4==4.6.0
peewee==3.6.4
pillow==5.2.0